fix: close phase Xb sprint X5 qa findings
