+++ crates/atm-daemon/src/peer_transport.rs
@@ -47,6 +47,7 @@ impl PeerTransportConfig {
             tracing::warn!(
                 subsystem = "peer_transport",
                 action = "config_default",
+                outcome = "default",
                 "no AtmConfig provided to PeerTransportConfig::from_config; using default remote_retry_budget"
             );
         }
@@ -105,6 +106,7 @@ fn daemon_peer_endpoint_from_env() -> Option<SocketAddr> {
             tracing::warn!(
                 subsystem = "peer_transport",
                 action = "env_parse",
+                outcome = "ignored",
                 "ignoring non-unicode ATM_DAEMON_PEER_ADDR value"
             );
             None
@@ -232,6 +234,9 @@ impl PeerClientTransport {
                     record.last_attempt_at = Some(IsoTimestamp::now());
                     record.last_error = Some(error.code.to_string());
                     tracing::warn!(
+                        subsystem = "peer_transport",
+                        action = "resume_replay",
+                        outcome = "skipped",
                         message_key = %record.message_key,
                         peer_addr = %record.peer_addr,
                         replay_attempt_count = record.attempt_count,
@@ -293,6 +298,9 @@ impl PeerClientTransport {
     fn persist_outcome_unknown_request(&self, request: &RequestEnvelope) -> Result<(), AtmError> {
         let Some((team, agent, message_key)) = replay_metadata_for_request(request) else {
             tracing::warn!(
+                subsystem = "peer_transport",
+                action = "persist",
+                outcome = "unknown",
                 request = ?request,
                 "remote delivery outcome is unknown but this request family does not support durable replay persistence",
             );
@@ -519,6 +527,9 @@ impl PeerClientTransport {
         let sleep_for =
             jittered_backoff(*retry_state.backoff, jitter_seed(endpoint, attempt)).min(remaining);
         tracing::warn!(
+            subsystem = "peer_transport",
+            action = "retry",
+            outcome = "retrying",
             peer_addr = %endpoint,
             attempt,
             sleep_ms = sleep_for.as_millis(),
@@ -668,6 +679,7 @@ fn wait_for_retry_backoff(terminate: &Arc<AtomicBool>, sleep_for: Duration) -> b
 }
 
 fn daemon_terminate_flag() -> Result<Arc<AtomicBool>, AtmError> {
+    // The lifecycle signal is cached globally because the handler requires one shared terminate flag, not per-transport construction wiring.
     static DAEMON_TERMINATE_FLAG: Mutex<Option<Arc<AtomicBool>>> = Mutex::new(None);
 
     let mut cached_flag = DAEMON_TERMINATE_FLAG.lock().map_err(|_| {
@@ -841,6 +853,9 @@ fn jitter_seed(endpoint: SocketAddr, attempt: u32) -> u64 {
         Ok(duration) => duration.as_nanos() as u64,
         Err(error) => {
             tracing::warn!(
+                subsystem = "peer_transport",
+                action = "jitter_seed",
+                outcome = "default",
                 %error,
                 "system clock is before the unix epoch; using deterministic jitter fallback"
             );
@@ -855,6 +870,9 @@ fn parse_peer_endpoint(raw: &str) -> Option<SocketAddr> {
         Ok(endpoint) => Some(endpoint),
         Err(error) => {
             tracing::warn!(
+                subsystem = "peer_transport",
+                action = "parse_endpoint",
+                outcome = "skipped",
                 %raw,
                 %error,
                 "parse_peer_endpoint: invalid address format"
